Move samples to packages.
This ensures that fully qualified class names are unique across samples.

According to the current README, we should additionally split these
files so that there is only one top-level class per file. Or,
alternatively, we could nest all classes inside a single top-level
class.

But see these related issues:
- https://github.com/jspecify/jspecify/issues/133
- https://github.com/jspecify/jspecify/issues/141
