+++ samples/nullnessUnspecifiedTypeParameter/nullnessunspecifiedtypeparameter/NullnessUnspecifiedTypeParameter.java
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+package nullnessunspecifiedtypeparameter;
+
 import org.jspecify.annotations.DefaultNonNull;
 import org.jspecify.annotations.Nullable;
